requirements.txt
django
django-heroku
gunicorn
requests
bs4
RPi.GPIO
whitenoise
pillow
